tests/wpt/web-platform-tests/css/work-in-progress/ttwf_bj/Along/ttwf-jstest-borderRadius.html

<!DOCTYPE html>
<html>
<head>
    <title>Test Background and border</title>
    <link rel="author" title="Jinlong Zhang" href="mailto:jinlongz@oupeng.com">
    <link rel="help" href="https://drafts.csswg.org/css3-background/#border-radius-tables">
    <link rel="help" href="">
    <meta name="flags" content="dom"><!--precondition-->
    <meta name="assert" content="Table with 'border-collapse: collapse;' declaration, it's 'border-radius' style will not work.">
    <style type="text/css">
        #testCon{border: 4px solid green; border-radius: 5px; border-collapse: collapse;}
        #testCon td{width: 60px; height: 60px; border: 4px solid green; border-radius: 10px;}
    </style>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <table id="testCon">
        <tr><td></td><td></td><td></td></tr>
        <tr><td></td><td></td><td></td></tr>
        <tr><td></td><td></td><td></td></tr>
    </table>
    <div id="log"></div>
    <script>
      test(function() {
        var testCon = document.getElementById('testCon');
        testCon.style.borderRadius = '5px';
        assert_equals(testCon.style.getPropertyValue("border-radius"),
                      "5px",
                      "Table's borderRadius is 5px, but borderRadius don't work.")
      }, "Table_borderRadius_borderCollapse");
    </script>
</body>
</html>
